refactor: update import paths for apmJson to match new casing
